Create TestAppLoader and simplify test code
